:sparkles: Fix: useInfiniteQuery에서 useQuery로 변겅
