+++ client/src/components/slide/GenreSlide.tsx
@@ -1,244 +1,236 @@
-import { useState, useEffect } from 'react';
-import { useInfiniteQuery } from '@tanstack/react-query';
-import { GetTVData, GetMovieData } from '../../api/api';
-import { ContentData } from '../../types/types';
-import ItemCard from '../ui/ItemCard';
-import SliderLoading from '../ui/exceptions/sliderLoading';
-import styled from 'styled-components';
-import SwiperCore, { Virtual, Navigation } from 'swiper';
-import { Swiper, SwiperSlide } from 'swiper/react';
-import 'swiper/css';
-import 'swiper/css/navigation';
+// import { useState, useEffect } from 'react';
+// import { useInfiniteQuery } from '@tanstack/react-query';
+// import { GetTVData, GetMovieData } from '../../api/api';
+// import { ContentData } from '../../types/types';
+// import ItemCard from '../ui/ItemCard';
+// import SliderLoading from '../ui/exceptions/sliderLoading';
+// import styled from 'styled-components';
+// import SwiperCore, { Virtual, Navigation } from 'swiper';
+// import { Swiper, SwiperSlide } from 'swiper/react';
+// import 'swiper/css';
+// import 'swiper/css/navigation';
 
-SwiperCore.use([Virtual, Navigation]);
+// SwiperCore.use([Virtual, Navigation]);
 
-const GenreSlide = ({ genre, path }: { genre: string, path: 'tv'|'movie' }) => {
-  const [, setSwiperRef] = useState<SwiperCore | null>(null);
-  const [size, setSize] = useState(getSize());
+// const breakpoints = {
+//   0: { slidesPerView: 3, slidesPerGroup: 3, spaceBetween: 10 },
+//   770: { slidesPerView: 4, slidesPerGroup: 4, spaceBetween: 14 },
+//   1024: { slidesPerView: 5, slidesPerGroup: 5, spaceBetween: 16 },
+//   1200: { slidesPerView: 6, slidesPerGroup: 6, spaceBetween: 18 }
+// };
 
-  function getSize() {
-    const width = window.innerWidth;
+// const GenreSlide = ({ genre, path }: { genre: string, path: 'tv'|'movie' }) => {
+//   const [, setSwiperRef] = useState<SwiperCore | null>(null);
+//   const [size, setSize] = useState(getSize());
 
-    if (width < 770) {
-      return 12;
-    } else if (width < 1024) {
-      return 16;
-    } else if (width < 1200) {
-      return 20;
-    } else {
-      return 24;
-    }
-  }
+//   function getSize() {
+//     const width = window.innerWidth;
 
-  const breakpoints = {
-    0: { slidesPerView: 3, slidesPerGroup: 3, spaceBetween: 10 },
-    770: { slidesPerView: 4, slidesPerGroup: 4, spaceBetween: 14 },
-    1024: { slidesPerView: 5, slidesPerGroup: 5, spaceBetween: 16 },
-    1200: { slidesPerView: 6, slidesPerGroup: 6, spaceBetween: 18 }
-  };
+//     if (width < 770) {
+//       return 12;
+//     } else if (width < 1024) {
+//       return 16;
+//     } else if (width < 1200) {
+//       return 20;
+//     } else {
+//       return 24;
+//     }
+//   }
 
-  const { data, fetchNextPage, hasNextPage, status } = useInfiniteQuery(
-    path === 'tv' ? ['tvGenreSlide', genre] : ['movieGenreSlide', genre],
-    ({ pageParam = 1 }) =>
-      path === 'tv'
-        ? GetTVData(genre, size, pageParam)
-        : GetMovieData(genre, size, pageParam),
-    {
-      getNextPageParam: (lastPage) => {
-        const currentPage = lastPage.currentPage;
-        const totalPages = lastPage.totalPages;
-        if (currentPage < totalPages) {
-          return currentPage + 1;
-        }
+//   const { data, fetchNextPage, hasNextPage, status } = useInfiniteQuery(
+//     path === 'tv' ? ['tvGenreSlide', genre] : ['movieGenreSlide', genre],
+//     ({ pageParam = 1 }) =>
+//       path === 'tv'
+//         ? GetTVData(genre, size, pageParam)
+//         : GetMovieData(genre, size, pageParam),
+//     {
+//       getNextPageParam: (lastPage) => {
+//         const currentPage = lastPage.currentPage;
+//         const totalPages = lastPage.totalPages;
+//         if (currentPage < totalPages) {
+//           return currentPage + 1;
+//         }
 
-        return undefined;
-      },
-    }
-  )
+//         return undefined;
+//       },
+//     }
+//   )
 
-  const handleNextPage = (e: React.MouseEvent) => {
-    const target = e.target as HTMLElement;
-    if (target.className === 'swiper-button-next') {
-      if (!hasNextPage) {
-        return
-      }
-      fetchNextPage()
-    }
-  }
+//   useEffect(() => {
+//     const handleResize = () => {
+//       setSize(getSize());
+//     };
+//     window.addEventListener('resize', handleResize);
 
-  useEffect(() => {
-    const handleResize = () => {
-      setSize(getSize());
-    };
-    window.addEventListener('resize', handleResize);
+//     return () => {
+//       window.removeEventListener('resize', handleResize);
+//     };
+//   }, []);
 
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
+//   if (status === 'loading') {
+//     return (
+//       <>
+//         <S_GenreTitle>{genre}</S_GenreTitle>
+//         <S_LoadingBox>
+//           <SliderLoading />
+//         </S_LoadingBox>
+//       </>
+//     )
+//   }
 
-  if (status === 'loading') {
-    return (
-      <>
-        <S_GenreTitle>{genre}</S_GenreTitle>
-        <S_LoadingBox>
-          <SliderLoading />
-        </S_LoadingBox>
-      </>
-    )
-  }
-
-  if (status === 'error') return <div>Error</div>;
+//   if (status === 'error') return <div>Error</div>;
   
-  if (status === 'success' && data.pages[0].content.length > 0) {
-    return (
-      <>
-        <S_GenreTitle>{genre}</S_GenreTitle>
-        <S_SwiperBox onClick={(e: React.MouseEvent) => handleNextPage(e)}>
-          <S_Swiper
-            onSwiper={setSwiperRef}
-            onReachEnd={() => {
-              if (hasNextPage) {
-                fetchNextPage();
-              }
-            }}
-            slidesPerView={6}
-            slidesPerGroup={5}
-            centeredSlides={false}
-            spaceBetween={18}
-            navigation={true}
-            watchOverflow={true}
-            breakpoints={breakpoints}
-            virtual
-          >
-            {data.pages.map((page) => (
-              <>
-                {page.content.map((item: ContentData) => (
-                  <S_SwiperSlide key={`${genre}-${item.id}`}>
-                    <ItemCard item={item} />
-                  </S_SwiperSlide>
-                ))}
-              </>
-            ))}
-          </S_Swiper>
-        </S_SwiperBox>
-      </>
-    );
-  }
-};
+//   if (status === 'success' && data.pages[0].content.length > 0) {
 
-export default GenreSlide;
+//     return (
+//       <>
+//         {console.log(data)};
+//         <S_GenreTitle>{genre}</S_GenreTitle>
+//         <S_SwiperBox>
+//           <S_Swiper
+//             onSwiper={setSwiperRef}
+//             onReachEnd={() => {
+//               if (hasNextPage) {
+//                 fetchNextPage();
+//               }
+//             }}
+//             slidesPerView={6}
+//             slidesPerGroup={5}
+//             centeredSlides={false}
+//             spaceBetween={18}
+//             navigation={true}
+//             watchOverflow={true}
+//             breakpoints={breakpoints}
+//             virtual
+//           >
+//             {data.pages.map((page) => (
+//               <>
+//                 {page.content.map((item: ContentData) => (
+//                   <S_SwiperSlide key={`${genre}-${item.id}`}>
+//                     <ItemCard item={item} />
+//                   </S_SwiperSlide>
+//                 ))}
+//               </>
+//             ))}
+//           </S_Swiper>
+//         </S_SwiperBox>
+//       </>
+//     );
+//   }
+// };
 
-const S_SwiperBox = styled.div`
-  position: relative;
-  overflow-x: hidden;
-  margin: 0;
-  padding: 0px 3.75rem;
-  width: 100%;
+// export default GenreSlide;
 
-  @media only screen and (max-width: 770px) {
-    padding: 0px 1.25rem;
-  }
-`;
+// const S_SwiperBox = styled.div`
+//   position: relative;
+//   overflow-x: hidden;
+//   margin: 0;
+//   padding: 0px 3.75rem;
+//   width: 100%;
 
-const S_Swiper = styled(Swiper)`
-  display: flex;
-  overflow: visible;
-  margin-top: 1.25rem;
-  margin-bottom: 3.75rem;
+//   @media only screen and (max-width: 770px) {
+//     padding: 0px 1.25rem;
+//   }
+// `;
 
-  @media only screen and (max-width: 770px) {
-    margin-top: 1rem;
-    margin-bottom: 2.5rem;
-  }
+// const S_Swiper = styled(Swiper)`
+//   display: flex;
+//   overflow: visible;
+//   margin-top: 1.25rem;
+//   margin-bottom: 3.75rem;
 
-  @media only screen and (max-width: 480px) {
-    margin-top: 0.75rem;
-    margin-bottom: 1rem;
-  }
+//   @media only screen and (max-width: 770px) {
+//     margin-top: 1rem;
+//     margin-bottom: 2.5rem;
+//   }
 
-  .swiper-button-prev,
-  .swiper-button-next {
-    top: 50%;
-    margin: 0;
-    padding: 1.5rem;
-    height: 100%;
-    transform: translateY(-55%);
-    background-repeat: no-repeat;
-    background: var(--color-bg-60);
-    color: var(--color-white-100);
-    text-shadow: 4px 4px 10px rgba(0, 0, 0, 0.4);
-    z-index: 10;
-    --swiper-navigation-size: 2.5rem;
-    opacity: 0;
+//   @media only screen and (max-width: 480px) {
+//     margin-top: 0.75rem;
+//     margin-bottom: 1rem;
+//   }
 
-    @media only screen and (max-width: 770px) {
-      padding: 1rem;
-      --swiper-navigation-size: 2rem;
-    }
+//   .swiper-button-prev,
+//   .swiper-button-next {
+//     top: 50%;
+//     margin: 0;
+//     padding: 1.5rem;
+//     height: 100%;
+//     transform: translateY(-55%);
+//     background-repeat: no-repeat;
+//     background: var(--color-bg-60);
+//     color: var(--color-white-100);
+//     text-shadow: 4px 4px 10px rgba(0, 0, 0, 0.4);
+//     z-index: 10;
+//     --swiper-navigation-size: 2.5rem;
+//     opacity: 0;
 
-    @media only screen and (max-width: 480px) {
-      padding: 0.8rem;
-      --swiper-navigation-size: 1.5rem;
-    }
-  }
+//     @media only screen and (max-width: 770px) {
+//       padding: 1rem;
+//       --swiper-navigation-size: 2rem;
+//     }
 
-  .swiper-button-prev {
-    left: -3.75rem;
+//     @media only screen and (max-width: 480px) {
+//       padding: 0.8rem;
+//       --swiper-navigation-size: 1.5rem;
+//     }
+//   }
 
-    @media only screen and (max-width: 770px) {
-      left: -1.25rem;
-    }
-  }
+//   .swiper-button-prev {
+//     left: -3.75rem;
 
-  .swiper-button-next {
-    right: -3.75rem;
+//     @media only screen and (max-width: 770px) {
+//       left: -1.25rem;
+//     }
+//   }
 
-    @media only screen and (max-width: 770px) {
-      right: -1.25rem;
-    }
-  }
+//   .swiper-button-next {
+//     right: -3.75rem;
 
-  .swiper-button-disabled {
-    display: none;
-  }
+//     @media only screen and (max-width: 770px) {
+//       right: -1.25rem;
+//     }
+//   }
 
-  &:hover {
-    .swiper-button-prev,
-    .swiper-button-next {
-      opacity: 1;
-      transition: opacity 0.3s ease;
-    }
-  }
-`;
+//   .swiper-button-disabled {
+//     display: none;
+//   }
 
-const S_SwiperSlide = styled(SwiperSlide)`
-  display: flex;
-  cursor: pointer;
-`;
+//   &:hover {
+//     .swiper-button-prev,
+//     .swiper-button-next {
+//       opacity: 1;
+//       transition: opacity 0.3s ease;
+//     }
+//   }
+// `;
 
-const S_LoadingBox = styled.div`
-  margin-top: 15px;
-  padding: 0px 3.75rem;
+// const S_SwiperSlide = styled(SwiperSlide)`
+//   display: flex;
+//   cursor: pointer;
+// `;
 
-  @media only screen and (max-width: 770px) {
-    padding: 0px 1.25rem;
-  }
+// const S_LoadingBox = styled.div`
+//   margin-top: 15px;
+//   padding: 0px 3.75rem;
 
-  @media only screen and (max-width: 480px) {
-    margin-top: 10px;
-  }
-`;
+//   @media only screen and (max-width: 770px) {
+//     padding: 0px 1.25rem;
+//   }
 
-const S_GenreTitle = styled.h2`
-  margin: 28px 0 5px 0;
-  padding: 0px 3.75rem;
-  color: var(--color-white-100);
-  font-size: 24px;
-  font-weight: 700;
+//   @media only screen and (max-width: 480px) {
+//     margin-top: 10px;
+//   }
+// `;
 
-  @media only screen and (max-width: 770px) {
-    margin: 25px 0 5px 0;
-    padding: 0px 1.25rem;
-  }
-`;
+// const S_GenreTitle = styled.h2`
+//   margin: 28px 0 5px 0;
+//   padding: 0px 3.75rem;
+//   color: var(--color-white-100);
+//   font-size: 24px;
+//   font-weight: 700;
+
+//   @media only screen and (max-width: 770px) {
+//     margin: 25px 0 5px 0;
+//     padding: 0px 1.25rem;
+//   }
+// `;
