+++ client/src/pages/Movie.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect, useRef } from 'react';
 import { useInView } from 'react-intersection-observer';
 import BannerSlide from '../components/slide/BannerSlide';
-import GenreSlide from '../components/slide/GenreSlide';
+// import GenreSlide from '../components/slide/GenreSlide';
+import GenreSlide from '../components/slide/GenreSlider';
 import styled from 'styled-components';
 import ListBtns from '../components/ui/ListBtns';
 import { scrollToTop } from '../utils/scrollToTop';
@@ -12,6 +13,56 @@ import bannerMovieImg3 from '../assets/banner_image/카운트.webp';
 import bannerMovieImg4 from '../assets/banner_image/암살.webp';
 import bannerMovieImg5 from '../assets/banner_image/기생충.webp';
 
+const bannerMovieImgs: BannerImgsType = [
+  {
+    name: bannerMovieImg1,
+    alt: '아바타 물의 길',
+    id: 109
+  },
+  {
+    name: bannerMovieImg2,
+    alt: '코코',
+    id: 326
+  },
+  {
+    name: bannerMovieImg3,
+    alt: '카운트',
+    id: 654
+  },
+  {
+    name: bannerMovieImg4,
+    alt: '암살',
+    id: 475
+  },
+  {
+    name: bannerMovieImg5,
+    alt: '기생충',
+    id: 201
+  },
+];
+
+const genres: string[] = [
+  '액션',
+  '드라마',
+  'SF',
+  '스릴러',
+  '애니메이션',
+  '코미디',
+  '가족',
+  '판타지',
+  '로맨스',
+  '공포',
+  '범죄',
+  '스포츠',
+  '음악',
+  'Made in Europe',
+  'Reality TV',
+  '역사',
+  '다큐멘터리',
+  '전쟁',
+  '서부'
+];
+
 function Movie() {
   const [visibleGenres, setVisibleGenres] = useState<Array<string>>([]);
   const currentIndex = useRef(4);
@@ -19,56 +70,6 @@ function Movie() {
     threshold: 0.1,
     triggerOnce: false,
   });
-
-  const bannerMovieImgs: BannerImgsType = [
-    {
-      name: bannerMovieImg1,
-      alt: '아바타 물의 길',
-      id: 109
-    },
-    {
-      name: bannerMovieImg2,
-      alt: '코코',
-      id: 326
-    },
-    {
-      name: bannerMovieImg3,
-      alt: '카운트',
-      id: 654
-    },
-    {
-      name: bannerMovieImg4,
-      alt: '암살',
-      id: 475
-    },
-    {
-      name: bannerMovieImg5,
-      alt: '기생충',
-      id: 201
-    },
-  ];
-
-  const genres: string[] = [
-    '액션',
-    '드라마',
-    'SF',
-    '스릴러',
-    '애니메이션',
-    '코미디',
-    '가족',
-    '판타지',
-    '로맨스',
-    '공포',
-    '범죄',
-    '스포츠',
-    '음악',
-    'Made in Europe',
-    'Reality TV',
-    '역사',
-    '다큐멘터리',
-    '전쟁',
-    '서부'
-  ];
 
   useEffect(() => {
     const genreSlice = genres.slice(0, currentIndex.current);
